Loading gathering items from CSV

+++ Arrowgene.Ddon.GameServer/Handler/ItemUseBagItemHandler.cs
@@ -46,7 +46,7 @@ namespace Arrowgene.Ddon.GameServer.Handler
             ntcData0.ItemList.ItemId = item.ItemId;
             ntcData0.ItemList.ItemNum = itemNum;
             ntcData0.ItemList.Unk3 = item.Unk3;
-            ntcData0.ItemList.StorageType = (byte) DestinationStorageType;
+            ntcData0.ItemList.StorageType = DestinationStorageType;
             ntcData0.ItemList.SlotNo = slotNo;
             ntcData0.ItemList.Color = item.Color; // ?
             ntcData0.ItemList.PlusValue = item.PlusValue; // ?
